Address several issues.
* changed firetotem and sentry to use the faster targetting method (closes #28)
* let invasion spawns respawn when there's a dead player on top (closes #20)
* medic cannot revive gladiators (cannot reproduce, seems to work fine)
* "was destroyed by an <unknown>" messages fixed
* Replace "chick" in GetMonsterKindString (fixes #18)
* Fixed max count when reviving things as a medic (fixes #17)
* Techs now drop at the beginning of sprees (fixes #29)
* weapon runes can now only spawn up to level 5 mods (needs testing #30)
* Make the streak go to the player owner of the attacker instead of the attacker (fixes #19)
* fix third slot being ignored when picking up runes as well as 4th slot being applied and hardmax being 15. (fixes #13)
* Give CONTENTS_MONSTERCLIP | MASK_SOLID to runes (closes #32)
* Add CONTENTS_MONSTERCLIP to MASK_BOTSOLIDX. Should make monsters unable to be tossed over monsterclips.
* Changed intro message to reflect the Revival branding.
* Restored weapons to base values. HG range addon is now 5.
* Free ammo regen removed.
* Medics will no longer spawn in invasion easy mode.
* Gone through some of QW's improvements.
* Split up magic.c a bit. Moved some #defines out into the lua as a consequence.
* Armory now properly lands the menu cursor on the "Next" line. (Addr. #24)

+++ src/combat/abilities/forcewall.c
@@ -265,7 +265,7 @@ void SpawnForcewall(edict_t *player, int type)
 	tr.endpos[2]++;
 	VectorCopy(tr.endpos, wall->s.origin); // copy origin to ending position of trace
   
-	if(fabs(forward[0]) > fabs(forward[1]))
+	if(fabsf(forward[0]) > fabsf(forward[1]))
 	{
 		// set wall thickness
 		wall->pos1[0] = wall->pos2[0] = wall->s.origin[0];
